+++ javax-kt/src/main/java/me/panpf/javaxkt/net/Netx.kt
@@ -30,3 +30,7 @@ inline fun CharSequence?.isIPV4(): Boolean = Netx.isIPV4(this)
  */
 inline fun CharSequence?.isIPV6(): Boolean = Netx.isIPV6(this)
 
+inline fun Long.ipLongToString(): String = Netx.ipLongToString(this)
+
+inline fun String.ipStringToLong(): Long = Netx.ipStringToLong(this)
+