+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## v0.8.2
 
 * :sparkles: Classx support class name
+* :sparkles: Netx adds ipStringToLong, ipLongToString method
 
 ## v0.8.1